pyproject.toml
[project]
name = "sipin-sip-delivery-registrator"
version = "0.1.1"
description = "Listens to deliveries of incoming SIP packages and registers the deliveries in a database."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "viaa-chassis",
    "pulsar-client",
    "psycopg>=3.2.9,<4.0.0",
    "psycopg-pool>=3.2.6,<4.0.0",
    "meemoo-cloudevents==0.1.0-rc.4",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
    "testcontainers>=4.12.0,<5.0.0"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--rootdir=./app/ -v"
testpaths = ["tests"]
